Add Menu for paletteChooser control button

+++ EmojiArt/EmojiArtModel.swift
@@ -16,9 +16,9 @@ struct EmojiArtModel: Encodable, Decodable {
         var x: Int  // offset from center
         var y: Int  // offset from center
         var size: Int
-        let id: UInt16
+        let id: Int
         
-        fileprivate init(text: String, x: Int, y: Int, size: Int, id: UInt16) {
+        fileprivate init(text: String, x: Int, y: Int, size: Int, id: Int) {
             self.text = text
             self.x = x
             self.y = y
@@ -45,6 +45,6 @@ struct EmojiArtModel: Encodable, Decodable {
     private var uniqueEmojiId = 0
     mutating func addEmoji(_ text: String, at location: (x: Int, y: Int), size: Int) {
         uniqueEmojiId += 1
-        emojis.append(Emoji(text: text, x: location.x, y: location.y, size: size, id: UInt16(uniqueEmojiId)))
+        emojis.append(Emoji(text: text, x: location.x, y: location.y, size: size, id: uniqueEmojiId))
     }
 }
